+++ spec/features/all_projects/export/export_users_can_see_an_export_landing_page_spec.rb
@@ -26,7 +26,8 @@ RSpec.feature "Export users can see the exports landing page" do
     user = create(:user, team: :business_support)
 
     sign_in_with_user(user)
-    click_on "Exports"
+    click_link "All projects"
+    click_link "Exports"
 
     expect(page).to have_content("funding agreement letter contacts, RPA and start-up grants")
     expect(page).to have_content("pre-opening grants for schools becoming academies")
@@ -38,7 +39,8 @@ RSpec.feature "Export users can see the exports landing page" do
     user = create(:user, team: :service_support)
 
     sign_in_with_user(user)
-    click_on "Exports"
+    click_link "All projects"
+    click_link "Exports"
 
     expect(page).to have_content("funding agreement letter contacts, RPA and start-up grants")
     expect(page).to have_content("pre-opening grants for schools becoming academies")